requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.0.10
django-cors-headers==4.3.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
idna==3.11
openpyxl==3.1.5
pillow==10.4.0
PyJWT==2.11.0
python-dotenv==1.0.1
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.7.0
